requirements.txt
pandas==2.2.0
numpy==1.26.4
tensorflow==2.15.0
scikit-learn==1.4.0
MetaTrader5==5.0.45
matplotlib==3.8.2
yfinance==0.2.36
python-dotenv==1.0.1
